Add country and county as per suggestion.
I'm using your instructions on How To Convert Township/Range/Section data to Latitude/Longitude Coordinates Using GeoLocate from the Arctos Handbook. You say to include the country and state. I've been including the county as well. Isn't it possible that there might be two locations in the same state with different meridians and baselines but the same TRS numbers?